Address PR review feedback for IBufferWriter<char>, #1315
- Throw ArgumentOutOfRangeException (instead of ArgumentException) for
  negative count/sizeHint in OpenStringBuilder and CharTermAttributeImpl,
  and update the corresponding doc comments and test assertions.
- Reorder the Advance bounds check to put count on the left for clarity.
- Convert OpenStringBuilder.AsSpan()/AsMemory() from instance methods to an
  OpenStringBuilderExtensions class (same namespace) with the full overload
  set (start, start/length, Index, Range) and null-to-default handling,
  modeled on Lucene.Net.MemoryExtensions.
- Move the IBufferWriter test files under Support/ (namespaces unchanged).
