+++ src/Lucene.Net.Tests/Support/Analysis/TokenAttributes/TestCharTermAttributeImplBufferWriter.cs
@@ -11,7 +11,7 @@ using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace Lucene.Net.Analysis.Util
+namespace Lucene.Net.Analysis.TokenAttributes
 {
     /*
      * Licensed to the Apache Software Foundation (ASF) under one or more
@@ -31,14 +31,16 @@ namespace Lucene.Net.Analysis.Util
      */
 
     /// <summary>
-    /// Tests for <see cref="OpenStringBuilder"/>'s <see cref="IBufferWriter{T}"/> implementation.
+    /// Tests for the <see cref="CharTermAttribute"/> class' <see cref="IBufferWriter{T}"/> implementation.
     /// </summary>
     [TestFixture]
     [LuceneNetSpecific]
-    public class TestOpenStringBuilderBufferWriter : LuceneTestCase
+    public class TestCharTermAttributeImplBufferWriter : LuceneTestCase
     {
         private const char DefaultChar = (char)0;
-        private const int DefaultBufferSize = 32; // copied from hardcoded value in OpenStringBuilder
+
+        // LUCENENET specific: 10 is MIN_BUFFER_SIZE, but it gets oversized
+        private static readonly int DefaultBufferSize = ArrayUtil.Oversize(10, RamUsageEstimator.NUM_BYTES_CHAR);
 
         // LUCENENET specific: works around a .NET Framework x64 RyuJIT bug where
         // ReadOnlyMemory<char>.Span returns an empty span when the property is read
@@ -59,46 +61,12 @@ namespace Lucene.Net.Analysis.Util
         private static char MemoryElementAt(ReadOnlyMemory<char> memory, int index)
             => memory.Span[index];
 
+        // LUCENENET specific: note that Clear behaves more like a Reset, which resets the position without clearing the
+        // buffer. so this test looks a lot more like the ResetWrittenCount in ArrayBufferWriter tests, than Clear.
         [Test]
-        public void ArrayBufferWriter_Ctor()
+        public void Clear()
         {
-            {
-                var output = new OpenStringBuilder();
-                Assert.AreEqual(DefaultBufferSize, output.FreeCapacity);
-                Assert.AreEqual(DefaultBufferSize, output.Capacity);
-                Assert.AreEqual(0, output.Length);
-                Assert.True(ReadOnlySpan<char>.Empty.SequenceEqual(output.AsSpan()));
-                Assert.True(MemoryEqualsMemory(ReadOnlyMemory<char>.Empty, output.AsMemory()));
-            }
-
-            {
-                var output = new OpenStringBuilder(200);
-                Assert.True(output.FreeCapacity >= 200);
-                Assert.True(output.Capacity >= 200);
-                Assert.AreEqual(0, output.Length);
-                Assert.True(ReadOnlySpan<char>.Empty.SequenceEqual(output.AsSpan()));
-                Assert.True(MemoryEqualsMemory(ReadOnlyMemory<char>.Empty, output.AsMemory()));
-            }
-
-            {
-                OpenStringBuilder output = default;
-                Assert.Null(output);
-            }
-        }
-
-        [Test]
-        public void Invalid_Ctor()
-        {
-            // LUCENENET specific - changed to ArgumentOutOfRangeException
-            Assert.Throws<ArgumentOutOfRangeException>(() => new OpenStringBuilder(0));
-            Assert.Throws<ArgumentOutOfRangeException>(() => new OpenStringBuilder(-1));
-            Assert.Throws<OutOfMemoryException>(() => new OpenStringBuilder(int.MaxValue));
-        }
-
-        [Test]
-        public void Reset()
-        {
-            var output = new OpenStringBuilder(256);
+            var output = new CharTermAttribute();
             int previousAvailable = output.FreeCapacity;
             WriteData(output, 2);
             Assert.True(output.FreeCapacity < previousAvailable);
@@ -113,7 +81,36 @@ namespace Lucene.Net.Analysis.Util
             char t1 = transientSpan[1];
             Assert.AreNotEqual(DefaultChar, t0);
             Assert.AreNotEqual(DefaultChar, t1);
-            output.Reset();
+            output.Clear();
+            Assert.AreEqual(t0, MemoryElementAt(transientMemory, 0));
+            Assert.AreEqual(t1, transientSpan[1]);
+
+            Assert.AreEqual(0, output.Length);
+            Assert.True(ReadOnlySpan<char>.Empty.SequenceEqual(output.AsSpan()));
+            Assert.True(MemoryEqualsMemory(ReadOnlyMemory<char>.Empty, output.AsMemory()));
+            Assert.AreEqual(previousAvailable, output.FreeCapacity);
+        }
+
+        // LUCENENET: this is equivalent to the ResetWrittenCount test
+        [Test]
+        public void SetLengthToZero()
+        {
+            var output = new CharTermAttribute();
+            int previousAvailable = output.FreeCapacity;
+            WriteData(output, 2);
+            Assert.True(output.FreeCapacity < previousAvailable);
+            Assert.True(output.Length > 0);
+            Assert.False(ReadOnlySpan<char>.Empty.SequenceEqual(output.AsSpan()));
+            Assert.False(MemoryEqualsMemory(ReadOnlyMemory<char>.Empty, output.AsMemory()));
+            Assert.True(SpanEqualsMemory(output.AsSpan(), output.AsMemory()));
+
+            ReadOnlyMemory<char> transientMemory = output.AsMemory();
+            ReadOnlySpan<char> transientSpan = output.AsSpan();
+            char t0 = MemoryElementAt(transientMemory, 0);
+            char t1 = transientSpan[1];
+            Assert.AreNotEqual(DefaultChar, t0);
+            Assert.AreNotEqual(DefaultChar, t1);
+            output.Length = 0;
             Assert.AreEqual(t0, MemoryElementAt(transientMemory, 0));
             Assert.AreEqual(t1, transientSpan[1]);
 
@@ -127,7 +124,7 @@ namespace Lucene.Net.Analysis.Util
         public void Advance()
         {
             {
-                var output = new OpenStringBuilder();
+                var output = new CharTermAttribute();
                 int capacity = output.Capacity;
                 Assert.AreEqual(capacity, output.FreeCapacity);
                 output.Advance(output.FreeCapacity);
@@ -136,7 +133,7 @@ namespace Lucene.Net.Analysis.Util
             }
 
             {
-                var output = new OpenStringBuilder();
+                var output = new CharTermAttribute();
                 output.Advance(output.Capacity);
                 Assert.AreEqual(output.Capacity, output.Length);
                 Assert.AreEqual(0, output.FreeCapacity);
@@ -146,7 +143,7 @@ namespace Lucene.Net.Analysis.Util
             }
 
             {
-                var output = new OpenStringBuilder(256);
+                var output = new CharTermAttribute();
                 WriteData(output, 2);
                 ReadOnlyMemory<char> previousMemory = output.AsMemory();
                 ReadOnlySpan<char> previousSpan = output.AsSpan();
@@ -158,7 +155,7 @@ namespace Lucene.Net.Analysis.Util
             }
 
             {
-                var output = new OpenStringBuilder();
+                var output = new CharTermAttribute();
                 _ = output.GetSpan(20);
                 WriteData(output, 10);
                 ReadOnlyMemory<char> previousMemory = output.AsMemory();
@@ -175,7 +172,7 @@ namespace Lucene.Net.Analysis.Util
         [Test]
         public void AdvanceZero()
         {
-            var output = new OpenStringBuilder();
+            var output = new CharTermAttribute();
             WriteData(output, 2);
             Assert.AreEqual(2, output.Length);
             ReadOnlyMemory<char> previousMemory = output.AsMemory();
@@ -192,13 +189,13 @@ namespace Lucene.Net.Analysis.Util
         public void InvalidAdvance()
         {
             {
-                var output = new OpenStringBuilder();
-                Assert.Throws<ArgumentException>(() => output.Advance(-1));
+                var output = new CharTermAttribute();
+                Assert.Throws<ArgumentOutOfRangeException>(() => output.Advance(-1));
                 Assert.Throws<InvalidOperationException>(() => output.Advance(output.Capacity + 1));
             }
 
             {
-                var output = new OpenStringBuilder();
+                var output = new CharTermAttribute();
                 WriteData(output, 100);
                 Assert.Throws<InvalidOperationException>(() => output.Advance(output.FreeCapacity + 1));
             }
@@ -207,7 +204,7 @@ namespace Lucene.Net.Analysis.Util
         [Test]
         public void GetSpan_DefaultCtor()
         {
-            var output = new OpenStringBuilder();
+            var output = new CharTermAttribute();
             Span<char> span = output.GetSpan();
             Assert.AreEqual(DefaultBufferSize, span.Length);
         }
@@ -216,42 +213,16 @@ namespace Lucene.Net.Analysis.Util
         [TestCaseSource(nameof(SizeHints))]
         public void GetSpan_DefaultCtor_WithSizeHint(int sizeHint)
         {
-            var output = new OpenStringBuilder();
+            var output = new CharTermAttribute();
             Span<char> span = output.GetSpan(sizeHint);
-            Assert.AreEqual(sizeHint <= DefaultBufferSize ? DefaultBufferSize : sizeHint, span.Length);
-        }
-
-        [Test]
-        public void GetSpan_InitSizeCtor()
-        {
-            var output = new OpenStringBuilder(100);
-            Span<char> span = output.GetSpan();
-            Assert.AreEqual(100, span.Length);
-        }
-
-        [Test]
-        [TestCaseSource(nameof(SizeHints))]
-        public void GetSpan_InitSizeCtor_WithSizeHint(int sizeHint)
-        {
-            {
-                var output = new OpenStringBuilder(256);
-                Span<char> span = output.GetSpan(sizeHint);
-                // LUCENENET specific: due to our oversize logic, changed to GreaterOrEqual and removed sizeHint addend
-                Assert.GreaterOrEqual(span.Length, sizeHint <= 256 ? 256 : sizeHint);
-            }
-
-            {
-                var output = new OpenStringBuilder(1000);
-                Span<char> span = output.GetSpan(sizeHint);
-                // LUCENENET specific: due to our oversize logic, changed to GreaterOrEqual and removed sizeHint addend
-                Assert.GreaterOrEqual(span.Length, sizeHint <= 1000 ? 1000 : sizeHint);
-            }
+            // LUCENENET specific: due to our oversize logic, changed to GreaterOrEqual
+            Assert.GreaterOrEqual(span.Length, sizeHint <= DefaultBufferSize ? DefaultBufferSize : sizeHint);
         }
 
         [Test]
         public void GetMemory_DefaultCtor()
         {
-            var output = new OpenStringBuilder();
+            var output = new CharTermAttribute();
             Memory<char> memory = output.GetMemory();
             Assert.AreEqual(DefaultBufferSize, memory.Length);
         }
@@ -260,54 +231,28 @@ namespace Lucene.Net.Analysis.Util
         [TestCaseSource(nameof(SizeHints))]
         public void GetMemory_DefaultCtor_WithSizeHint(int sizeHint)
         {
-            var output = new OpenStringBuilder();
+            var output = new CharTermAttribute();
             Memory<char> memory = output.GetMemory(sizeHint);
-            Assert.AreEqual(sizeHint <= DefaultBufferSize ? DefaultBufferSize : sizeHint, memory.Length);
+            // LUCENENET specific: due to our oversize logic, changed to GreaterOrEqual
+            Assert.GreaterOrEqual(memory.Length, sizeHint <= DefaultBufferSize ? DefaultBufferSize : sizeHint);
         }
 
         [Test]
         public void GetMemory_ExceedMaximumBufferSize_WithSmallStartingSize()
         {
-            var output = new OpenStringBuilder(256);
+            var output = new CharTermAttribute();
             Assert.Throws<OutOfMemoryException>(() => output.GetMemory(int.MaxValue));
         }
 
-        [Test]
-        public void GetMemory_InitSizeCtor()
-        {
-            var output = new OpenStringBuilder(100);
-            Memory<char> memory = output.GetMemory();
-            Assert.AreEqual(100, memory.Length);
-        }
-
-        [Test]
-        [TestCaseSource(nameof(SizeHints))]
-        public void GetMemory_InitSizeCtor_WithSizeHint(int sizeHint)
-        {
-            {
-                var output = new OpenStringBuilder(256);
-                Memory<char> memory = output.GetMemory(sizeHint);
-                // LUCENENET specific: due to our oversize logic, changed to GreaterOrEqual and removed sizeHint addend
-                Assert.GreaterOrEqual(memory.Length, sizeHint <= 256 ? 256 : sizeHint);
-            }
-
-            {
-                var output = new OpenStringBuilder(1000);
-                Memory<char> memory = output.GetMemory(sizeHint);
-                // LUCENENET specific: due to our oversize logic, changed to GreaterOrEqual and removed sizeHint addend
-                Assert.GreaterOrEqual(memory.Length, sizeHint <= 1000 ? 1000 : sizeHint);
-            }
-        }
-
         // LUCENENET specific: This test allocates a very large buffer to verify that Advance()
-        // performs its bounds check purely arithmetically (m_len vs. m_buf.Length); Advance() itself
-        // never touches buffer memory, and WriteData only touches the first 1,000 chars. If the
-        // allocation fails with OutOfMemoryException, the test is a no-op. Marked [Slow] due to the
-        // large allocation.
+        // performs its bounds check purely arithmetically (termLength vs. termBuffer.Length);
+        // Advance() itself never touches buffer memory, and WriteData only touches the first 1,000
+        // chars. If the allocation fails with OutOfMemoryException, the test is a no-op. Marked
+        // [Slow] due to the large allocation.
         //
         // NOTE: The upstream ArrayBufferWriter test gated this on Windows/macOS, and so do we. The
         // risk is real on Linux but comes from the allocation itself, not from Advance(): under memory
-        // overcommit the kernel may grant new char[2_000_000_000] optimistically, then the GC's
+        // overcommit the kernel may grant the large buffer optimistically, then the GC's
         // zero-initialization of the array touches every page, which can invoke the OOM killer (an
         // uncatchable SIGKILL, not an OutOfMemoryException) and tear down the whole test run.
         [Test]
@@ -321,10 +266,21 @@ namespace Lucene.Net.Analysis.Util
             try
             {
                 {
-                    var output = new OpenStringBuilder(2_000_000_000);
+                    var output = new CharTermAttribute();
+
+                    // LUCENENET specific: we don't (currently) have a ctor that takes a capacity,
+                    // so we first request a span to force the buffer to grow to at least this size.
+                    // The size stays under ~1.9B so it does not hit the ArrayUtil.Oversize int
+                    // overflow clamp; the actual buffer is still oversized past the request.
+                    // (The returned span is discarded; the grown buffer is retained by output.)
+                    _ = output.GetSpan(1_500_000_000);
+
                     WriteData(output, 1_000);
+
+                    // LUCENENET specific: the buffer is oversized past the requested size, so the
+                    // over-advance boundary is Capacity-relative rather than a fixed request-based value.
                     Assert.Throws<InvalidOperationException>(() => output.Advance(int.MaxValue));
-                    Assert.Throws<InvalidOperationException>(() => output.Advance(2_000_000_000 - 1_000 + 1));
+                    Assert.Throws<InvalidOperationException>(() => output.Advance(output.FreeCapacity + 1));
                 }
             }
             catch (OutOfMemoryException) { }
@@ -334,7 +290,7 @@ namespace Lucene.Net.Analysis.Util
         public void GetMemoryAndSpan()
         {
             {
-                var output = new OpenStringBuilder();
+                var output = new CharTermAttribute();
                 WriteData(output, 2);
                 Span<char> span = output.GetSpan();
                 Memory<char> memory = output.GetMemory();
@@ -350,7 +306,7 @@ namespace Lucene.Net.Analysis.Util
             }
 
             {
-                var output = new OpenStringBuilder();
+                var output = new CharTermAttribute();
                 WriteData(output, 2);
                 ReadOnlyMemory<char> writtenSoFarMemory = output.AsMemory();
                 ReadOnlySpan<char> writtenSoFar = output.AsSpan();
@@ -387,9 +343,9 @@ namespace Lucene.Net.Analysis.Util
         }
 
         [Test]
-        public void GetSpanShouldAtleastDoubleWhenGrowing()
+        public void GetSpanShouldAtLeastDoubleWhenGrowing()
         {
-            var output = new OpenStringBuilder(256);
+            var output = new CharTermAttribute();
             WriteData(output, 100);
             int previousAvailable = output.FreeCapacity;
 
@@ -404,7 +360,7 @@ namespace Lucene.Net.Analysis.Util
         public void GetSpanOnlyGrowsAboveThreshold()
         {
             {
-                var output = new OpenStringBuilder();
+                var output = new CharTermAttribute();
                 _ = output.GetSpan();
                 int previousAvailable = output.FreeCapacity;
 
@@ -416,7 +372,7 @@ namespace Lucene.Net.Analysis.Util
             }
 
             {
-                var output = new OpenStringBuilder();
+                var output = new CharTermAttribute();
                 _ = output.GetSpan(10);
                 int previousAvailable = output.FreeCapacity;
 
@@ -431,31 +387,32 @@ namespace Lucene.Net.Analysis.Util
         [Test]
         public void InvalidGetMemoryAndSpan()
         {
-            var output = new OpenStringBuilder();
+            var output = new CharTermAttribute();
             WriteData(output, 2);
-            Assert.Throws<ArgumentException>(() => output.GetSpan(-1));
-            Assert.Throws<ArgumentException>(() => output.GetMemory(-1));
+            Assert.Throws<ArgumentOutOfRangeException>(() => output.GetSpan(-1));
+            Assert.Throws<ArgumentOutOfRangeException>(() => output.GetMemory(-1));
         }
 
         [Test]
         public void MultipleCallsToGetSpan()
         {
-            var output = new OpenStringBuilder(300);
+            var output = new CharTermAttribute();
             Assert.True(MemoryMarshal.TryGetArray(output.GetMemory(), out ArraySegment<char> array));
             GCHandle pinnedArray = GCHandle.Alloc(array.Array, GCHandleType.Pinned);
             try
             {
                 int previousAvailable = output.FreeCapacity;
-                Assert.True(previousAvailable >= 300);
-                Assert.True(output.Capacity >= 300);
+                Assert.True(previousAvailable >= DefaultBufferSize);
+                Assert.True(output.Capacity >= DefaultBufferSize);
                 Assert.AreEqual(previousAvailable, output.Capacity);
                 Span<char> span = output.GetSpan();
                 Assert.True(span.Length >= previousAvailable);
-                Assert.True(span.Length >= 256);
+                Assert.True(span.Length >= DefaultBufferSize);
                 Span<char> newSpan = output.GetSpan();
                 Assert.AreEqual(span.Length, newSpan.Length);
                 // LUCENENET specific: changed expected to IntPtr.Zero due to presumed Xunit vs. NUnit differences
-                Assert.AreEqual(IntPtr.Zero, Unsafe.ByteOffset(ref MemoryMarshal.GetReference(span), ref MemoryMarshal.GetReference(newSpan)));
+                Assert.AreEqual(IntPtr.Zero,
+                    Unsafe.ByteOffset(ref MemoryMarshal.GetReference(span), ref MemoryMarshal.GetReference(newSpan)));
                 Assert.AreEqual(span.Length, output.GetSpan().Length);
             }
             finally
